Set loid on the client
