+++ src/lib/randomString.es6.js
@@ -1,0 +1,12 @@
+function randomString(len) {
+  var id = [];
+  var chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+
+  for (var i = 0; i < len; i++) {
+    id.push(chars.charAt(Math.floor(Math.random() * chars.length)));
+  }
+
+  return id.join('');
+}
+
+export default randomString;
